# Conflicts:
#	database/migrations/tenant/2022_09_28_145600_addcollumn_product_options_table.php

+++ app/Http/Controllers/Tenant/PageController.php
@@ -62,13 +62,13 @@ class PageController extends Controller
         if (!$page->save()) {
             return back()->withInput()->with("error", "Error creating page.");
         }
-
+       
         URL::create([
-            'url' => 'pagina/'.$page->url,
+            'url' => 'page/'.$page->url,
             'entity' => 'PAGE',
             'entity_id' => $page->id,
         ])->save();
-
+       
         return redirect()->route('tenant.pages.index')->with("success", "Page created successfully!");
     }
 
@@ -108,8 +108,10 @@ class PageController extends Controller
             $page->delete();
 
             $url = URL::firstWhere('url', $page->url);
-
-            $url->delete();
+            if($url!=null){
+                $url->delete();
+            }
+               
 
             DB::commit();
         } catch (QueryException $e) {
